refactor(ai): Keep AI functionality closer
